Rename the WriteToFile thing just ExportFormat
